requirements.txt
ansible==6.7.0
ansible-core==2.13.13
cffi==1.17.1
cryptography==45.0.4
jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
pycparser==2.22
python-dotenv==1.0.1
PyYAML==6.0.2
resolvelib==0.8.1
